requirements.txt
requests==2.31.0
urllib3==2.0.4
xmltodict==0.13.0
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
